GS - geofence - allow remote layers/dynamic styles
By default set to false upstream, need to set it to True in the
geOrchestra default context.

Note: there is an unexpected behaviour with the properties resolutions
in GeoFence though (multiple beans for properties resolution, which
clashes one another, and actually use the unexpected properties - the
default ones defined in the xml).